Admin routes mm

+++ back-end/app/controllers/documents_controller.rb
@@ -8,14 +8,22 @@ class DocumentsController < ApplicationController
         end
      end
       def destroy
-        doc = Document.find(params[:id])
-        if doc
+        token = request.headers['Authentication'].split(' ')[1]
+        payload = decode(token) 
+        user = User.find(payload['user_id'])
+        if user.is_admin
+          doc = Document.find(params[:id])
+          if doc
             doc.destroy
             render json: doc
-        else
+          else
             render json: {message: 'Failed to destroy Document'}
+          end
+        else
+        render json: {message: "Unauthorized Route"}
         end
       end
+    private
       def doc_params
         params.permit(:name, :order_id,:file)
       end
